chore: renames MenuList
